\name{TO1}
\alias{TO1}
\docType{data}
\title{
Adjudicated Toronto Youth Data (Sample 1)
}
\description{
 Number of criminal unique court contacts for 378 individuals in Toronto,
 Ontario, Canada.
}
\usage{data(TO1)}
\format{
  A 378 by 31 matrix where each row contains the number of
  unique court contacts per year for an individual from the age of 8 to 38.
}
\keyword{datasets}
