\name{TO1adj}
\alias{TO1adj}
\docType{data}
\title{
Adjusted Adjudicated Toronto Youth Data (Sample 1)
}
\description{
 Number of unique court contacts for 378 individuals in Toronto,
 Ontario, Canada adjusted for time-at-risk by the "divide and round" approach.  
}
\usage{data(TO1adj)}
\format{
  A 378 by 31 matrix where each row contains the number of
  unique court contacts per year corrected for time-at-risk for an individual from the age of 8 to 38.
}
\keyword{datasets}
