\name{TO1Risk}
\alias{TO1Risk}
\docType{data}
\title{
Time-at-risk for the Adjudicated Toronto Youth Data (Sample 1)
}
\description{
 Time-at-risk per year for 378 individuals in the Toronto,
 Ontario, Canada.
}
\usage{data(TO1Risk)}
\format{
  A 378 by 31 matrix where each row contains the time-at-risk
   per year for an individual from the age of 8 to 38.
}
\keyword{datasets}
