\name{TO2}
\alias{TO2}
\docType{data}
\title{
Adjudicated Toronto Youth Data (Sample 2)
}
\description{
 Number of criminal offenses for 386 individuals in Toronto,
 Ontario, Canada.
}
\usage{data(TO2)}
\format{
  A 386 by 30 matrix where each row contains the number of
  criminal offenses per year for an individual from the age of 9 to 38.
}
\keyword{datasets}
