\name{TO2adj}
\alias{TO2adj}
\docType{data}
\title{
Adjusted Adjudicated Toronto Youth Data (Sample 2)
}
\description{
 Number of unique court contacts for 386 individuals in Toronto,
 Ontario, Canada adjusted for time-at-risk by the "divide and round" approach.  
}
\usage{data(TO2adj)}
\format{
  A 386 by 30 matrix where each row contains the number of
  unique court contacts per year corrected for time-at-risk for an individual from the age of 9 to 38.
}
\keyword{datasets}
