\name{TO2Risk}
\alias{TO2Risk}
\docType{data}
\title{
Time-at-risk for the Adjudicated Toronto Youth Data (Sample 2)
}
\description{
 Time-at-risk for 386 individuals in the Toronto,
 Ontario, Canada.
}
\usage{data(TO2Risk)}
\format{
  A 386 by 30 matrix where each row contains the number of
  criminal offenses per year for an individual from the age of 9 to 38.
}
\keyword{datasets}
